+++ spec/lib/concierge/suppliers/ciirus/commands/reservations_fetcher_spec.rb
@@ -57,6 +57,7 @@ RSpec.describe Ciirus::Commands::ReservationsFetcher do
 
         expect(result).not_to be_success
         expect(result.error.code).to eq :savon_error
+        expect(result.error.data).to be_nil
       end
     end