+++ spec/lib/concierge/suppliers/rentals_united/commands/booking_spec.rb
@@ -46,6 +46,9 @@ RSpec.describe RentalsUnited::Commands::Booking do
     result = subject.call
     expect(result).not_to be_success
     expect(result.error.code).to eq("1")
+    expect(result.error.data).to eq(
+      "Property is not available for a given dates"
+    )
 
     event = Concierge.context.events.last.to_h
     expect(event[:message]).to eq(
@@ -56,6 +59,8 @@ RSpec.describe RentalsUnited::Commands::Booking do
   end
 
   context "when response from the api is not well-formed xml" do
+    let(:expected_error_message) { "Error response could not be recognised (no `Status` tag in the response)" }
+
     it "returns a result with an appropriate error" do
       stub_data = read_fixture("rentals_united/bad_xml.xml")
       stub_call(:post, credentials.url) { [200, {}, stub_data] }
@@ -64,11 +69,10 @@ RSpec.describe RentalsUnited::Commands::Booking do
 
       expect(result).not_to be_success
       expect(result.error.code).to eq(:unrecognised_response)
+      expect(result.error.data).to eq(expected_error_message)
 
       event = Concierge.context.events.last.to_h
-      expect(event[:message]).to eq(
-        "Error response could not be recognised (no `Status` tag in the response)"
-      )
+      expect(event[:message]).to eq(expected_error_message)
       expect(event[:backtrace]).to be_kind_of(Array)
       expect(event[:backtrace].any?).to be true
     end